Move k8s code into own file
